+++ bin/refresh-mining-pool-configs.js
@@ -1,4 +1,5 @@
 #!/usr/bin/env node
+"use strict";
 var os = require('os');
 var path = require('path');
 var dotenv = require("dotenv");